lesson20_drf/library/requirements.txt
Django>=4.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=23.3
drf-yasg==1.21.7
python-decouple==3.8